Update controller.sh
add economic check = 0, disables the check

Fix condition,
The code provided contains an error in the for loop when checking the condition within the charging_conditions array.

The problem:
The charging_conditions array contains strings and not the actual conditions. When you iterate through the array and use ((condition)), Bash checks that the value of condition (not the content it points to) is not null, which is always true since it is a string.