Update victron-venus-os-install.sh
add cp sample.config.txt to config.txt

This option will not overwrite the destination_file if it already exists. There is no output or error message if the file is not copied, the command simply exits silently. If the target file does not exist, it will be created as usual.